mailreader/requirements.txt
python-dotenv
flask
psycopg2-binary
nltk
pandas
stanfordcorenlp
sklearn
scikit-learn